% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathogenDose.R
\name{PathogenDose}
\alias{PathogenDose}
\title{Probability of illness from exposure to wastewater}
\usage{
PathogenDose(doseresp = "exp", alpha, beta = 0, Pillinf,
  path_sewage_dist = "lunif", p_alpha, p_beta, dose)
}
\arguments{
\item{doseresp}{The dose response model being used. Type \code{"exp"} for single 
parameter exponential, \code{"bp"} for two parameter beta-poisson, or \code{"f1"} for 
the two parameter hypergeometric1F1, commonly used for norovirus}

\item{alpha}{First parameter for the dose response}

\item{beta}{Second parameter for the dose response. Leave blank if using single 
paramter exponential dose response}

\item{Pillinf}{The fraction of infections that become illnesses 
If Pillinf is a distribution not a single value, use the function Fraction to 
generate a monte carlo table. Can be a single value or a monte carlo vector 
generated from the \code{frac} function. P _ill/inf The probability of illness 
if infected. Fraction in decimal form. i.e. for norovirus, 60 percent of infections 
become illnesses, so \code{frac=0.6} (McBride et al. 2013)}

\item{path_sewage_dist}{The distribution of the pathogen in sewage. 
Choose either \code{'lunif'} for a uniform distribution or \code{'lnorm'} for a 
normal distribution}

\item{p_alpha}{The first parameter value for the distribution of pathogen in 
sewage. For \code{'lunif'} alpha_sew is the min concentration in log 10 copies/L, 
for \code{'lnorm'} it is the mean value in log 10 copies/L.}

\item{p_beta}{The second parameter value for the distribution of pathogen in 
sewage. For \code{'lunif'} beta_sew is the max concentration in log 10 copies/L, 
for \code{'lnorm'} it is the standard deviation in log 10 units.}

\item{dose}{Returned dose from Wastewater Dose function. Defaults to WWdose}
}
\value{
Probability of illness \code{pill1}. Vector contains \code{count} number 
of samplings. To calculate illness from \code{n} number of pathogens, use the formula 
totpill=1-((1-pill1)(1-pill2)(1-pill3)(1-pill4)...)
}
\description{
Calculates the probability of illness for a specific pathogen based 
the amount of wastewater consumed during the swimming event
}
\details{
This function computes the probability of illness from exposure to wastewater
}
\examples{
\dontrun{
}
}
