% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WastewaterDose.R
\name{WastewaterDose}
\alias{WastewaterDose}
\title{Dose of Wastewater}
\usage{
WastewaterDose(indic_enviro_conc, indic_sewage_dist = "lunif", i_alpha,
  i_beta, seed = 123, count = 10000)
}
\arguments{
\item{indic_enviro_conc}{The concentration of the indicator in environmental 
waters-the dependent variable in log10 Copies/L}

\item{indic_sewage_dist}{Which distribution, log uniform: \code{'lunif'}  or 
log normal: \code{'lnorm'} the indicator concentration in sewage follows.}

\item{i_alpha}{First value for distrubution, if lunif, alpha is the min concentration, 
for lnorm it is the alpha value.}

\item{i_beta}{Second value for distribution, if lunif, beta is the max concentration, 
for lnorm it is the beta value.}

\item{seed}{Optional paramter. Sets the seed of R's random number generator, which 
is useful for creating simulations that can be reproduced. Default is 123}

\item{count}{Optional parameter. How many Monte Carlo samplings. Default is 10,000}
}
\value{
Wastewater dose, \code{WWdose}, in Liters. Vector contains  \code{count} 
number of samplings.
}
\description{
Calculate how much wastewater a swimmer is exposed to based on a single 
indicator concentration value. Water dose is from "Water ingestion during swimming 
activities in a pool: a pilot study", lognormal distribution with a mean of 2.92 
and a standard deviation of 1.43 (Dufor et al. 2006).
}
\details{
This function calculates how much wastewater a swimmer is exposed to based on a 
single indicator concentration value
}
\examples{
\dontrun{
}
}
